remarks_commit

+++ tasks/kulik_a_the_most_different_adjacent/common/include/common.hpp
@@ -9,7 +9,7 @@
 namespace kulik_a_the_most_different_adjacent {
 
 using InType = std::vector<double>;
-using OutType = std::pair<int, int>;
+using OutType = std::pair<uint64_t, uint64_t>;
 using TestType = std::string;
 using BaseTask = ppc::task::Task<InType, OutType>;
